Move private `ClassMap` type to `class` module.
